requirements.txt
solders
solana
jupiter-python-sdk
streamlit
pandas
python-dotenv
requests
httpx
aiohttp
plotly
watchdog
